+++ server/routes/prisons/sessionTemplates/index.ts
@@ -1,0 +1,22 @@
+import { RequestHandler, Router } from 'express'
+import { Services } from '../../../services'
+import asyncMiddleware from '../../../middleware/asyncMiddleware'
+import SessionTemplatesController from './sessionTemplatesController'
+import SingleSessionTemplateController from './singleSessionTemplateController'
+
+export default function routes(services: Services): Router {
+  const router = Router()
+
+  const get = (path: string | string[], handler: RequestHandler) => router.get(path, asyncMiddleware(handler))
+
+  const sessionTemplates = new SessionTemplatesController(services.prisonService, services.sessionTemplateService)
+  const singleSessionTemplate = new SingleSessionTemplateController(
+    services.prisonService,
+    services.sessionTemplateService,
+  )
+
+  get('/prisons/:prisonId([A-Z]{3})/session-templates', sessionTemplates.view())
+  get('/prisons/:prisonId/session-templates/:reference', singleSessionTemplate.view())
+
+  return router
+}
