+++ server/routes/testutils/appSetup.ts
@@ -18,7 +18,6 @@ import cookieSession from 'cookie-session'
 import createError from 'http-errors'
 
 import routes from '../index'
-import prisonsRoutes from '../prisons/prisons'
 
 import nunjucksSetup from '../../utils/nunjucksSetup'
 import errorHandler from '../../errorHandler'
@@ -54,10 +53,7 @@ function appSetup(services: Services, production: boolean, userSupplier: () => E
   })
   app.use(express.json())
   app.use(express.urlencoded({ extended: true }))
-
-  app.use('/', routes(services))
-  app.use('/prisons', prisonsRoutes(services))
-
+  app.use(routes(services))
   app.use((req, res, next) => next(createError(404, 'Not found')))
   app.use(errorHandler(production))
